+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# BetterBlizzFrames 1.9.5j
+## All versions
+### Bugfix
+- Fix issue with custom font addition validation causing fonts to not be added.
+## The Burning Crusade
+### New
+- Add "Improved Addon List" setting to TBC as well (Misc).
+### Tweak
+- Update absorb stuff to work on TBC. The accuracy of it is TBD.
+## All Classics
+### New
+- Add right-click options for "Bigger Healthbars" setting allowing you to disable it on individual frames.
+
 # BetterBlizzFrames 1.9.5i
 ## Midnight
 ### Tweak